+++ site/components/module/navigation.tsx
@@ -1,5 +1,5 @@
 import { link } from "@site/components/element/link";
-import { component, element, faSvgIconStore, registerComponent, style } from "qrilljs/core";
+import { component, element, faSvgIconStore, registerStyle, style } from "qrilljs/core";
 import type { HComponentFn, HSvgBrandsIconName, Store } from "qrilljs/core";
 import { F_XLARGE, S_XLARGE } from "qrilljs/core";
 
@@ -32,7 +32,7 @@ export function navigation(store: Store): HComponentFn<NavigationArgument> {
         }),
     ];
 
-    registerComponent(store, Navigation, component_styles);
+    registerStyle(store, Navigation, component_styles);
 
     return component(Navigation, ({ navitem }) => (
         <Navigation>