+++ site/pages/tags/[tag].html.tsx
@@ -15,7 +15,7 @@ export function rootPageFnParameters(): RootParameter[] {
 
 export default function Root(store: Store): HRootPageFn<RootParameter> {
     const Page = page(store);
-    const PageMainArea = element(store, "main", { name: "page-main-area" });
+    const PageMainArea = element(store, { tag: "main", name: "page-main-area" });
     const Summaries = summaries(store);
 
     const styles = [